refactor: break up `main`
The `main.rs` file is too busy, so let's split some of it out and allow
each of the files to be dedicated to a specific piece of work more.

This change:
* Moves the bootstrap code to `src/persistence/bootstrap.rs`
* Moves the endpoints and router creation to `src/router.rs`
* Updates all the references and imports
